+++ packages/manager/src/components/RegionSelect/RegionSelect.utils.tsx
@@ -9,7 +9,7 @@ import type {
   RegionFilterValue,
 } from './RegionSelect.types';
 import type { AccountAvailability, Capabilities, Region } from '@linode/api-v4';
-import type { LinodeCreateType } from 'src/features/Linodes/LinodesCreate/types';
+import type { LinodeCreateType } from 'src/features/Linodes/LinodeCreatev2/types';
 
 const NORTH_AMERICA = CONTINENT_CODE_TO_CONTINENT.NA;
