+++ packages/manager/src/features/Linodes/LinodesCreate/ApiAwarenessModal/CurlTabPanel.tsx
@@ -11,7 +11,7 @@ import { useIsAkamaiAccount } from 'src/hooks/useIsAkamaiAccount';
 import { sendApiAwarenessClickEvent } from 'src/utilities/analytics/customEventAnalytics';
 import { generateCurlCommand } from 'src/utilities/codesnippets/generate-cURL';
 
-import { CodeBlock } from '../CodeBlock/CodeBlock';
+import { CodeBlock } from './CodeBlock/CodeBlock';
 
 import type { CreateLinodeRequest } from '@linode/api-v4/lib/linodes';
