+++ packages/manager/src/features/Linodes/LinodesCreate/ApiAwarenessModal/SDKTabPanel.tsx
@@ -7,7 +7,7 @@ import { LD_DX_TOOLS_METRICS_KEYS } from 'src/constants';
 import { generateGoLinodeSnippet } from 'src/utilities/codesnippets/generate-goSDKSnippet';
 import { generatePythonLinodeSnippet } from 'src/utilities/codesnippets/generate-pythonSDKSnippet';
 
-import { CodeBlock } from '../CodeBlock/CodeBlock';
+import { CodeBlock } from './CodeBlock/CodeBlock';
 import { GoSDKResources } from './GoSDKResources';
 import { PythonSDKResources } from './PythonSDKResources';
