--- packages/manager/src/features/Linodes/LinodesCreate/constants.ts
@@ -1,5 +1,0 @@
-export const CROSS_DATA_CENTER_CLONE_WARNING =
-  'Cloning a powered off instance across data centers may cause long periods of down time.';
-
-export const REGION_CAVEAT_HELPER_TEXT =
-  'A Linode may be assigned only to a VPC in the same region.';
