+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeRebuild/ImageEmptyState.tsx
