+++ packages/manager/src/features/Linodes/types.ts
@@ -1,4 +1,4 @@
-import type { LinodeCreateType } from 'src/features/Linodes/LinodesCreate/types';
+import type { LinodeCreateType } from './LinodeCreatev2/types';
 import type { BaseQueryParams } from 'src/utilities/queryParams';
 
 export type DialogType =
