+++ packages/manager/src/features/StackScripts/UserDefinedFieldsPanel/AppInfo.tsx
